finished reworking initial components
