+++ transformer_lens/components/layer_norm_pre.py
@@ -1,0 +1,55 @@
+"""Hooked Transformer Components.
+
+This module contains all the components (e.g. :class:`Attention`, :class:`MLP`, :class:`LayerNorm`)
+needed to create many different types of generative language models. They are used by
+:class:`transformer_lens.HookedTransformer`.
+"""
+from jaxtyping import Float
+import torch
+import torch.nn as nn
+from transformer_lens.hook_points import HookPoint
+from transformer_lens.HookedTransformerConfig import HookedTransformerConfig
+from typing import Dict, Union
+
+
+# LayerNormPre
+# I fold the LayerNorm weights and biases into later weights and biases.
+# This is just the 'center and normalise' part of LayerNorm
+# Centering is equivalent to just deleting one direction of residual space,
+# and is equivalent to centering the weight matrices of everything writing to the residual stream
+# Normalising is a funkier non-linear operation, that projects the residual stream onto the unit hypersphere
+class LayerNormPre(nn.Module):
+    def __init__(self, cfg: Union[Dict, HookedTransformerConfig]):
+        """LayerNormPre - the 'center and normalise' part of LayerNorm. Length is
+        normally d_model, but is d_mlp for softmax. Not needed as a parameter. This
+        should only be used in inference mode after folding in LayerNorm weights"""
+        super().__init__()
+        if isinstance(cfg, Dict):
+            cfg = HookedTransformerConfig.from_dict(cfg)
+        self.cfg = cfg
+        self.eps = self.cfg.eps
+
+        # Adds a hook point for the normalisation scale factor
+        self.hook_scale = HookPoint()  # [batch, pos]
+        # Hook Normalized captures LN output - here it's a vector with std 1 and mean 0
+        self.hook_normalized = HookPoint()  # [batch, pos, length]
+
+    def forward(
+        self,
+        x: Union[
+            Float[torch.Tensor, "batch pos d_model"],
+            Float[torch.Tensor, "batch pos head_index d_model"],
+        ],
+    ) -> Union[
+        Float[torch.Tensor, "batch pos d_model"],
+        Float[torch.Tensor, "batch pos head_index d_model"],
+    ]:
+        if self.cfg.dtype not in [torch.float32, torch.float64]:
+            x = x.to(torch.float32)
+
+        x = x - x.mean(axis=-1, keepdim=True)  # [batch, pos, length]
+        scale: Union[
+            Float[torch.Tensor, "batch pos 1"],
+            Float[torch.Tensor, "batch pos head_index 1"],
+        ] = self.hook_scale((x.pow(2).mean(-1, keepdim=True) + self.eps).sqrt())
+        return self.hook_normalized(x / scale).to(self.cfg.dtype)
