+++ transformer_lens/components/token_typed_embed.py
@@ -1,0 +1,32 @@
+"""Hooked Transformer Components.
+
+This module contains all the components (e.g. :class:`Attention`, :class:`MLP`, :class:`LayerNorm`)
+needed to create many different types of generative language models. They are used by
+:class:`transformer_lens.HookedTransformer`.
+"""
+
+from jaxtyping import Int
+import torch
+import torch.nn as nn
+from transformer_lens.HookedTransformerConfig import HookedTransformerConfig
+from typing import Dict, Union
+
+
+class TokenTypeEmbed(nn.Module):
+    """
+    The token-type embed is a binary ids indicating whether a token belongs to sequence A or B. For example, for two sentences: "[CLS] Sentence A [SEP] Sentence B [SEP]", token_type_ids would be [0, 0, ..., 0, 1, ..., 1, 1]. `0` represents tokens from Sentence A, `1` from Sentence B. If not provided, BERT assumes a single sequence input. Typically, shape is (batch_size, sequence_length).
+
+    See the BERT paper for more information: https://arxiv.org/pdf/1810.04805.pdf
+    """
+
+    def __init__(self, cfg: Union[Dict, HookedTransformerConfig]):
+        super().__init__()
+        if isinstance(cfg, Dict):
+            cfg = HookedTransformerConfig.from_dict(cfg)
+        self.cfg = cfg
+        self.W_token_type = nn.Parameter(
+            torch.empty(2, self.cfg.d_model, dtype=cfg.dtype)
+        )
+
+    def forward(self, token_type_ids: Int[torch.Tensor, "batch pos"]):
+        return self.W_token_type[token_type_ids, :]
